Rename contexttimer to timer
This keeps our names nice and parallel:

```
from .bloom import BloomFilter
from .cache import redis_cache
from .hyper import HyperLogLog
from .nextid import NextId
from .redlock import Redlock
from .timer import ContextTimer
```
